refactor(13710): Refactor PointRangeWeight
